requirements.txt
Django==1.10.5
django-extensions==1.7.6

djangorestframework==3.5.4
django-filter==1.0.1
django-crispy-forms==1.6.1
django-guardian==1.4.6
coreapi==2.2.4
Markdown==2.6.8

bpython==0.16